LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-convolvernode-interface/convolver-response-4-chan-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.

Harness Error (FAIL), message = Unhandled rejection: The object is in an invalid state.

PASS # AUDIT TASK RUNNER STARTED. 
PASS Executing "initialize" 
FAIL Executing "1-channel input" promise_test: Unhandled rejection with value: object "IndexSizeError: The index is not in the allowed range."
FAIL Executing "2-channel input" promise_test: Unhandled rejection with value: object "IndexSizeError: The index is not in the allowed range."
FAIL Executing "3-channel input" promise_test: Unhandled rejection with value: object "IndexSizeError: The index is not in the allowed range."
FAIL Executing "4-channel input" promise_test: Unhandled rejection with value: object "IndexSizeError: The index is not in the allowed range."
FAIL Executing "5.1-channel input" promise_test: Unhandled rejection with value: object "IndexSizeError: The index is not in the allowed range."
PASS Executing "delayed buffer set" 
FAIL Executing "count 1, 2-channel in" promise_test: Unhandled rejection with value: object "IndexSizeError: The index is not in the allowed range."
FAIL Executing "count 1, 4-channel in" promise_test: Unhandled rejection with value: object "IndexSizeError: The index is not in the allowed range."
FAIL Executing "count 1, 5.1-channel in" promise_test: Unhandled rejection with value: object "IndexSizeError: The index is not in the allowed range."
PASS Audit report 
PASS > [initialize] Convolver response with one channel 
PASS   new AudioBuffer({numberOfChannels: 2, length: 4, sampleRate: 8192}) did not throw an exception. 
PASS < [initialize] All assertions passed. (total 1 assertions) 
PASS > [1-channel input] produces 2-channel output 
PASS > [2-channel input] produces 2-channel output 
PASS > [3-channel input] 3->2 downmix producing 2-channel output 
PASS > [4-channel input] 4->2 downmix producing 2-channel output 
PASS > [5.1-channel input] 5.1->2 downmix producing 2-channel output 
PASS > [delayed buffer set] Delayed set of 4-channel response 
FAIL X Output with delayed setting of convolver buffer should have contain at least one value different from 0. assert_true: expected true got false
FAIL < [delayed buffer set] 1 out of 1 assertions were failed. assert_true: expected true got false
PASS > [count 1, 2-channel in] 2->1 downmix because channel count is 1 
PASS > [count 1, 4-channel in] 4->1 downmix because channel count is 1 
PASS > [count 1, 5.1-channel in] 5.1->1 downmix because channel count is 1 
FAIL # AUDIT TASK RUNNER FINISHED: 1 out of 10 tasks were failed. assert_true: expected true got false
